Added switch for LocalAccountTokenFilterPolicy

+++ examples/scripts/ConfigureRemotingForAnsible.ps1
@@ -1,69 +1,80 @@
+<#
+.SYNOPSIS 
+    Configure a Windows host for remote management with Ansible
+.DESCRIPTION 
+    This script checks the current WinRM (PS Remoting) configuration and makes
+    the necessary changes to allow Ansible to connect, authenticate and
+    execute PowerShell commands.
+
+    All events are logged to the Windows EventLog, useful for unattended runs.
+
+    Use option -Verbose in order to see the verbose output messages.
+.NOTES 
+    Written by Trond Hindenes <trond@hindenes.com>
+    Updated by Chris Church <cchurch@ansible.com>
+    Updated by Michael Crilly <mike@autologic.cm>
+    Updated by Anton Ouzounov <Anton.Ouzounov@careerbuilder.com>
+    Updated by Nicolas Simond <contact@nicolas-simond.com>
+    Updated by Dag Wieërs <dag@wieers.com>
+    Updated by Jordan Borean <jborean93@gmail.com>
+    Updated by Erwan Quélin <erwan.quelin@gmail.com>
+    Updated by David Norman <david@dkn.email>
+    Updated by Philipp Hoffmann <>
+    
+    Version 1.0 - 2014-07-06
+    Version 1.1 - 2014-11-11
+    Version 1.2 - 2015-05-15
+    Version 1.3 - 2016-04-04
+    Version 1.4 - 2017-01-05
+    Version 1.5 - 2017-02-09
+    Version 1.6 - 2017-04-18
+    Version 1.7 - 2017-11-23
+    Version 1.8 - 2018-02-23
+    Version 1.9 - 2018-09-21
+    Version 1.10 - 2020-07-31
+.LINK 
+    https://github.com/ansible/ansible/blob/devel/examples/scripts/ConfigureRemotingForAnsible.ps1
+.EXAMPLE 
+    TBD
+.PARAMETER SubjectName
+    Use option -SubjectName to specify the CN name of the certificate. This
+    defaults to the system's hostname and generally should not be specified.
+.PARAMETER CertValidityDays
+    Use option -CertValidityDays to specify how long this certificate is valid
+    starting from today. So you would specify -CertValidityDays 3650 to get
+    a 10-year valid certificate.
+.PARAMETER SkipNetworkProfileCheck
+    Use option -SkipNetworkProfileCheck to skip the network profile check.
+    Without specifying this the script will only run if the device's interfaces
+    are in DOMAIN or PRIVATE zones.  Provide this switch if you want to enable
+    WinRM on a device with an interface in PUBLIC zone.
+.PARAMETER ForceNewSSLCert
+    Use option -ForceNewSSLCert if the system has been SysPreped and a new
+    SSL Certificate must be forced on the WinRM Listener when re-running this
+    script. This is necessary when a new SID and CN name is created.
+.PARAMETER GlobalHttpFirewallAccess
+    No description provided.
+.PARAMETER DisableBasicAuth
+    Use option -DisableBasicAuth to disable basic authentication.
+.PARAMETER EnableCredSSP
+    Use option -EnableCredSSP to enable CredSSP as an authentication option.
+.PARAMETER EnableLocalAccountTokenFilterPolicy
+    Use option -EnableLocalAccountTokenFilterPolicy to enable the according
+    local GPO. This token is required if you are using a local user account
+    to connect via ansible. This setting is not required if you connect via
+    Active Directory user account.
+#>
 #Requires -Version 3.0
-
-# Configure a Windows host for remote management with Ansible
-# -----------------------------------------------------------
-#
-# This script checks the current WinRM (PS Remoting) configuration and makes
-# the necessary changes to allow Ansible to connect, authenticate and
-# execute PowerShell commands.
-#
-# All events are logged to the Windows EventLog, useful for unattended runs.
-#
-# Use option -Verbose in order to see the verbose output messages.
-#
-# Use option -CertValidityDays to specify how long this certificate is valid
-# starting from today. So you would specify -CertValidityDays 3650 to get
-# a 10-year valid certificate.
-#
-# Use option -ForceNewSSLCert if the system has been SysPreped and a new
-# SSL Certificate must be forced on the WinRM Listener when re-running this
-# script. This is necessary when a new SID and CN name is created.
-#
-# Use option -EnableCredSSP to enable CredSSP as an authentication option.
-#
-# Use option -DisableBasicAuth to disable basic authentication.
-#
-# Use option -SkipNetworkProfileCheck to skip the network profile check.
-# Without specifying this the script will only run if the device's interfaces
-# are in DOMAIN or PRIVATE zones.  Provide this switch if you want to enable
-# WinRM on a device with an interface in PUBLIC zone.
-#
-# Use option -SubjectName to specify the CN name of the certificate. This
-# defaults to the system's hostname and generally should not be specified.
-
-# Written by Trond Hindenes <trond@hindenes.com>
-# Updated by Chris Church <cchurch@ansible.com>
-# Updated by Michael Crilly <mike@autologic.cm>
-# Updated by Anton Ouzounov <Anton.Ouzounov@careerbuilder.com>
-# Updated by Nicolas Simond <contact@nicolas-simond.com>
-# Updated by Dag Wieërs <dag@wieers.com>
-# Updated by Jordan Borean <jborean93@gmail.com>
-# Updated by Erwan Quélin <erwan.quelin@gmail.com>
-# Updated by David Norman <david@dkn.email>
-#
-# Version 1.0 - 2014-07-06
-# Version 1.1 - 2014-11-11
-# Version 1.2 - 2015-05-15
-# Version 1.3 - 2016-04-04
-# Version 1.4 - 2017-01-05
-# Version 1.5 - 2017-02-09
-# Version 1.6 - 2017-04-18
-# Version 1.7 - 2017-11-23
-# Version 1.8 - 2018-02-23
-# Version 1.9 - 2018-09-21
-
-# Support -Verbose option
 [CmdletBinding()]
-
 Param (
     [string]$SubjectName = $env:COMPUTERNAME,
     [int]$CertValidityDays = 1095,
     [switch]$SkipNetworkProfileCheck,
-    $CreateSelfSignedCert = $true,
     [switch]$ForceNewSSLCert,
     [switch]$GlobalHttpFirewallAccess,
     [switch]$DisableBasicAuth = $false,
-    [switch]$EnableCredSSP
+    [switch]$EnableCredSSP,
+    [switch]$SkipEnableLocalAccountTokenFilterPolicy = $false
 )
 
 Function Write-Log
@@ -294,18 +305,20 @@ Else
     Write-Verbose "PS Remoting is already enabled."
 }
 
-# Ensure LocalAccountTokenFilterPolicy is set to 1
-# https://github.com/ansible/ansible/issues/42978
-$token_path = "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Policies\System"
-$token_prop_name = "LocalAccountTokenFilterPolicy"
-$token_key = Get-Item -Path $token_path
-$token_value = $token_key.GetValue($token_prop_name, $null)
-if ($token_value -ne 1) {
-    Write-Verbose "Setting LocalAccountTOkenFilterPolicy to 1"
-    if ($null -ne $token_value) {
-        Remove-ItemProperty -Path $token_path -Name $token_prop_name
-    }
-    New-ItemProperty -Path $token_path -Name $token_prop_name -Value 1 -PropertyType DWORD > $null
+if (!$SkipEnableLocalAccountTokenFilterPolicy) {
+  # Ensure LocalAccountTokenFilterPolicy is set to 1
+  # https://github.com/ansible/ansible/issues/42978
+  $token_path = "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Policies\System"
+  $token_prop_name = "LocalAccountTokenFilterPolicy"
+  $token_key = Get-Item -Path $token_path
+  $token_value = $token_key.GetValue($token_prop_name, $null)
+  if ($token_value -ne 1) {
+      Write-Verbose "Setting LocalAccountTOkenFilterPolicy to 1"
+      if ($null -ne $token_value) {
+          Remove-ItemProperty -Path $token_path -Name $token_prop_name
+      }
+      New-ItemProperty -Path $token_path -Name $token_prop_name -Value 1 -PropertyType DWORD > $null
+  }
 }
 
 # Make sure there is a SSL listener.
